fix: use Long for sample rate in PyTorch IValue.from() call
IValue.from() doesn't accept Int, only Long for integer values.
Convert SAMPLE_RATE to Long before passing to the model.

+++ app/src/main/java/com/looplingo/horizon/vad/SileroVadDetector.kt
@@ -188,9 +188,10 @@ class SileroVadDetector(private val context: Context) {
 
                 // Run inference: model(audio_chunk, sample_rate)
                 // The model maintains internal LSTM state automatically
+                // Note: IValue.from() requires Long, not Int, for integer values
                 val output = currentModule.forward(
                     IValue.from(inputTensor),
-                    IValue.from(SAMPLE_RATE)
+                    IValue.from(SAMPLE_RATE.toLong())
                 )
 
                 // Extract speech probability from output tensor [1, 1]
